Added the game logic for dealing damages

+++ src/Cyberpunk2078/Assets/Scripts/Projectile/IDamageable.cs
@@ -1,5 +1,5 @@
 ﻿
 public interface IDamageable
 {
-    int ApplyDamage(int rawDamage);
+    float ApplyDamage(float rawDamage);
 }
